+++ srvcheck/chains/__init__.py
@@ -7,5 +7,6 @@ from .tezos import Tezos
 from .solana import Solana
 from .aptos import Aptos
 from .near import Near
+from .amplitude import Amplitude
 
-CHAINS = [Substrate, Tendermint, Tezos, Lisk, Solana, Aptos, Near]
+CHAINS = [Substrate, Tendermint, Tezos, Lisk, Solana, Aptos, Near, Amplitude]
